Removed the variant site cache associated to each `SearchState`.
Previously, this was used to buffer a site/allele ID combination as a
read entered or exited a variant site, and was only committed to the
actual `variant_site_path` traversed upon read extension.

However the site/allele ID combination can be committed greedily to the
`variant_site_path` instead as the recording of this information is
currently guaranteed to occur always before read extension. This will
lose the `SearchState` if no match is found anyway.