Consistent production of reference files.
Some sanity checks:

* The base reference (first allele in each var site of prg) produced
using prg_regions_parser is exactly the same as that produced using
prg_local_parser

* Producing the inferred reference uses this base reference and the
inferred vcf, as before, but now the inferred vcf contains all sites.
The inferred fasta is not exactly the same as that inferred in `infer`
command, because we pick the first allele at het calls. In infer, we
pick the single most likely haploid (homozygous) call. But the sanity
check is that asking to produce inferred ref. taking the first allele at
each var site, we get the base reference.